+++ service/product/package.json
@@ -14,9 +14,11 @@
     "@babel/core": "^7.26.9",
     "@babel/preset-env": "^7.26.9",
     "@babel/preset-typescript": "^7.26.0",
+    "@types/jest": "^29.5.14",
     "@types/node": "^22.13.4",
     "babel-jest": "^29.7.0",
     "jest": "^29.7.0",
+    "ts-jest": "^29.2.6",
     "typescript": "^5.7.3"
   },
   "dependencies": {
